pyproject.toml
[tool.poetry]
name = "giordane_api"
version = "0.1.0"
description = ""
authors = ["dunossauro <mendesxeduardo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-sqlalchemy = "^2.4.4"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.24.1"
flask-migrate = "^2.5.3"
faker = "^5.0.0"
factory_boy = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
